# app/Http/Requests/ClientStoreUpdateRequest.php
<?php

namespace App\Http\Requests;

use Illuminate\Foundation\Http\FormRequest;

class ClientStoreUpdateRequest extends FormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            'name' => [
                'required',
                'string',
                'min:3',
                'max:60',
            ],
            'social_name' => [
                'required',
                'string',
                'min:3',
                'max:60',
            ],
            'birth_date' => [
                'required',
                'date_format:d/m/Y',
            ],
            'cpf' => [
                'required',
                'size:11',
                'string',
                "unique:clients,cpf,{$this->id},id",
            ],
            'photo' => [
                $this->isMethod('PUT') ? 'nullable' : 'required',
                'image',
                'mimes:jpeg,png,jpg,gif',
                'max:2048',
            ],
        ];
    }
}